fixtures fix #161
